Debug Cleanup
Reapply the debug cleanup code following the reverts. Turns out the issue was that the header file was defining a long string (to be used as a separator in the output). Simply removing that declaration and replacing all instances with the string constant fixed the issue. It also freed up some more memory.
Also adding back some of the previous functionality.